moved productCard
